requirements.txt
# -----------------------
# Core data stack
# -----------------------
pandas>=2.0
numpy>=1.24
pymongo>=4.6
yfinance>=0.2.40
PyYAML>=6.0
python-dotenv>=1.0
requests>=2.31

# -----------------------
# NLP / Sentiment
# -----------------------
vaderSentiment>=3.3.2
textblob>=0.18.0

# -----------------------
# Machine Learning
# -----------------------
scikit-learn>=1.3
joblib>=1.3
tensorflow>=2.15

# -----------------------
# Visualization / Dashboard
# -----------------------
streamlit>=1.30
matplotlib>=3.8

# -----------------------
# Big Data / MLOps (REQUIRED by brief)
# -----------------------
kafka-python>=2.0.2
pyspark>=3.5.0
mlflow>=2.10
dvc>=3.40

# Airflow (pin version for stability)
apache-airflow==2.10.5
